+++ tests/ChargeFormTest.php
@@ -1,9 +1,8 @@
 <?php
 
-
 use idarex\pingppyii2\ChargeForm;
 
-class ChargeFormTests extends PHPUnit_Framework_TestCase
+class ChargeFormTest extends PHPUnit_Framework_TestCase
 {
     public function testCreate()
     {
@@ -18,5 +17,18 @@ class ChargeFormTests extends PHPUnit_Framework_TestCase
         $properties = $reflectionClass->getDefaultProperties();
 
         $this->assertTrue(array_diff_key($chargeForm->getCharge(true), $properties) == []);
+    }
+
+    /**
+     * @depends testCreate
+     */
+    public function testGetWechatSignature()
+    {
+        $chargeForm = new ChargeForm();
+        $chargeForm->load(require 'data/wechat-charge.php', '');
+        $this->assertTrue($chargeForm->validate() && $chargeForm->create(), 'Create wechat charge');
+
+        $signature = $chargeForm->getWechatSignature(null, 'https://m.idarex.com');
+        $this->assertNotEmpty($signature, 'Signature can not be empty');
     }
 }