+++ tests/data/wechat-charge.php
@@ -1,0 +1,14 @@
+<?php
+
+return [
+    'order_no' => 'pingppyii2test' . (string)time(),
+    'amount' => '999999999',
+    'channel' => 'wx_pub',
+    'currency' => 'cny',
+    'client_ip' => '127.0.0.1',
+    'subject' => 'idarex pingpp-yii2 tests',
+    'body' => 'idarex pingpp-yii2 tests body',
+    'extra' => [
+        'open_id' => 'wx60ab301bf65da4b1',
+    ],
+];