+++ CHANGELOG.md
@@ -4,7 +4,10 @@ Pingpp Extension for Yii2 Change Log
 v1.1
 --------------------
 
-- Modify `$chargeForm->create()` Only returns *`true`* or *`false`* ! important
+- Modify `ChargeForm::create()` Only returns *`true`* or *`false`* ! important
 - Add `RedEnvelopeForm`
 - Add `redEnvelopeRetrieve`
 - Add `redEnvelopeList`
+- Add `ChargeForm::getWechatSignature()`
+
+closes: idarex/pingpp-yii2#10 idarex/pingpp-yii2#11 idarex/pingpp-yii2#12